refactor(graph): Refactoring the attribute line component to extract common logic into a separate file
